backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
supabase==2.7.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# HTTP
httpx==0.27.2
aiohttp==3.10.5

# Scraping
telethon==1.36.0
playwright==1.47.0
beautifulsoup4==4.12.3
lxml==5.3.0

# NLP / ML
spacy==3.7.6
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2

# Scheduling
apscheduler==3.10.4

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
geopy==2.4.1
